Этап 4 в работе

+++ src/main/java/com/example/bankcards/entity/Transaction.java
@@ -4,7 +4,6 @@ package com.example.bankcards.entity;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -29,7 +28,7 @@ public class Transaction {
     @NotNull
     @Positive(message = "Сумма транзакции должна быть положительной")
     @Column(name = "amount", nullable = false, precision = 19, scale = 2)
-    private BigDecimal amount;
+    private double amount;
 
     @NotNull
     @Column(name = "timestamp", nullable = false)
@@ -43,7 +42,7 @@ public class Transaction {
     public Transaction() {
     }
 
-    public Transaction(Card fromCard, Card toCard, BigDecimal amount, LocalDateTime timestamp, TransactionStatus status) {
+    public Transaction(Card fromCard, Card toCard, double amount, LocalDateTime timestamp, TransactionStatus status) {
         this.fromCard = fromCard;
         this.toCard = toCard;
         this.amount = amount;
@@ -75,11 +74,11 @@ public class Transaction {
         this.toCard = toCard;
     }
 
-    public BigDecimal getAmount() {
+    public double getAmount() {
         return amount;
     }
 
-    public void setAmount(BigDecimal amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
